refactor: Improve project structure, logging, and testing
This commit addresses a comprehensive code review by implementing several major improvements to the project's structure, maintainability, and professionalism.

Key changes include:

1.  **Backend Refactoring:** The monolithic `backend/main.py` has been split into a modular structure with a `core` directory containing `brain.py` and `toolbox.py`. The main file now only handles API and WebSocket routing.

2.  **Professional Logging:** All `print()` statements in the backend have been replaced with a standard Python `logging` setup, configured in `core/logger.py`.

3.  **Enhanced Security:** A Content-Security-Policy (CSP) meta tag has been added to the frontend to mitigate XSS risks.

4.  **Unit Testing:** A `tests` directory has been added with a `pytest` suite for the `Toolbox`'s file system functions, establishing a foundation for future testing.

5.  **Containerization:** A multi-stage `Dockerfile` has been added to allow for easy, consistent deployment of the backend service.

6.  **Gitignore:** A `.gitignore` file has been added to exclude generated files and artifacts from the repository, and previously tracked temporary files have been removed.
