+++ .gitignore
@@ -1,0 +1,31 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.pyc
+*.[oa]
+*.so
+
+# C extensions
+*.pyd
+*.dll
+
+# Logs and databases
+*.log
+*.db
+*.sqlite3
+
+# Environment directories
+.env
+.venv
+venv/
+env/
+
+# Test output
+.pytest_cache/
+.coverage
+htmlcov/
+
+# IDE / Editor specific
+.vscode/
+.idea/
+*.swp
+*.swo
