+++ frontend/index.html
@@ -3,6 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="Content-Security-Policy" content="default-src 'self'; style-src 'self'; script-src 'self'; connect-src 'self' ws://localhost:8000;">
     <title>الوكيل الذكي التفاعلي</title>
     <link rel="stylesheet" href="style.css">
 </head>
